Add delivery type to PerformanceResourceTiming.
This defines a concept of "delivery type" which currently describes the
difference between resources delivered from the HTTP cache and those not
(currently only detectable via transferSize). It is expected to be
extended to describe resources delivered from other stores, such as the
consuming resources from the HTML preload buffer and navigational
prefetches.